+++ app/src/main/AndroidManifest.xml
@@ -14,14 +14,14 @@
         android:screenOrientation="portrait"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".MainActivity">
+        <activity android:name=".main.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".ItemActivity"></activity>
+        <activity android:name=".item.ItemActivity"></activity>
         <activity android:name=".ImageviewerActivity"></activity>
         <activity android:name=".RecordingActivity"></activity>
     </application>
