+++ app/src/main/java/app/ddf/danskdatahistoriskforening/DatePickerFragment.java
@@ -5,13 +5,12 @@ import android.app.Dialog;
 import android.support.v4.app.DialogFragment;
 import android.os.Bundle;
 import android.widget.DatePicker;
-import android.widget.TextView;
 
-import java.text.DateFormat;
 import java.text.ParseException;
-import java.text.SimpleDateFormat;
 import java.util.Calendar;
 import java.util.Date;
+
+import app.ddf.danskdatahistoriskforening.item.RegisterDetailsFragment;
 
 public class DatePickerFragment extends DialogFragment implements DatePickerDialog.OnDateSetListener{
