+++ app/src/main/java/app/ddf/danskdatahistoriskforening/dal/IDAO.java
@@ -1,4 +1,4 @@
-package app.ddf.danskdatahistoriskforening;
+package app.ddf.danskdatahistoriskforening.dal;
 
 import android.content.Context;
 import android.net.Uri;
