+++ app/src/main/java/app/ddf/danskdatahistoriskforening/main/FrontFragment.java
@@ -1,4 +1,4 @@
-package app.ddf.danskdatahistoriskforening;
+package app.ddf.danskdatahistoriskforening.main;
 
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
@@ -6,6 +6,8 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Button;
+
+import app.ddf.danskdatahistoriskforening.R;
 
 public class FrontFragment extends Fragment implements View.OnClickListener{
