+++ app/src/main/res/layout/fragment_front.xml
@@ -2,7 +2,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="app.ddf.danskdatahistoriskforening.FrontFragment">
+    tools:context=".main.FrontFragment">
 
 
     <LinearLayout
