+++ app/src/main/res/layout/fragment_item_list.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="app.ddf.danskdatahistoriskforening.ItemListFragment">
+    tools:context=".main.ItemListFragment">
 
     <ListView
         android:layout_width="wrap_content"
